+++ python/src/indian_fakedata/__init__.py
@@ -39,6 +39,9 @@ from indian_fakedata.utils.exporter import format_profiles, save_profiles
 from indian_fakedata.utils.appearance import generate_appearance, get_region
 from indian_fakedata.utils.employment import generate_employment_timeline
 from indian_fakedata.utils.skills import generate_skills
+from indian_fakedata.utils.transliterate import (
+    transliterate, script_for_language, contains_indic,
+)
 from indian_fakedata.utils.schema import get_profile_schema, validate_profile
 from indian_fakedata.utils.privacy import strip_pii, PII_FIELDS
 
@@ -73,6 +76,9 @@ __all__ = [
     "get_region",
     "generate_employment_timeline",
     "generate_skills",
+    "transliterate",
+    "script_for_language",
+    "contains_indic",
     "get_profile_schema",
     "validate_profile",
     "strip_pii",